Refactor: Move Plotly graph builders to separate file
